Move the IR interpreter out into its own file. Rename it in the UI.  First CMake and Android fixes.
