+++ Core/MIPS/IR/IRInterpreter.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "Common/CommonTypes.h"
+
+class MIPSState;
+struct IRInst;
+
+u32 IRInterpret(MIPSState *mips, const IRInst *inst, const u32 *constPool, int count);
