+++ UI/GameSettingsScreen.cpp
@@ -1060,7 +1060,7 @@ void DeveloperToolsScreen::CreateViews() {
 	}
 #endif
 
-	static const char *cpuCores[] = { "Interpreter", "Dynarec (JIT)", "IRJit" };
+	static const char *cpuCores[] = { "Interpreter", "Dynarec (JIT)", "IR Interpreter" };
 	PopupMultiChoice *core = list->Add(new PopupMultiChoice(&g_Config.iCpuCore, gr->T("CPU Core"), cpuCores, 0, ARRAY_SIZE(cpuCores), sy->GetName(), screenManager()));
 	if (!canUseJit) {
 		core->HideChoice(1);
